% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lda.R
\name{coef.lda}
\alias{coef.lda}
\alias{coef.qda}
\title{Extract Coefficients from Discriminant Analysis}
\usage{
\method{coef}{lda}(object, ...)

\method{coef}{qda}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link[MASS:lda]{MASS::lda()}} or \code{\link[MASS:qda]{MASS::qda()}}}

\item{...}{Other arguments (ignored)}
}
\value{
Coefficients (\code{scaling} components) extracted from the model \code{object}.
}
\description{
Provides \code{coef} methods to extract the coefficients (\code{scaling} components) for variables from discriminant analysis methods \code{\link[MASS:lda]{MASS::lda()}} and \code{\link[MASS:qda]{MASS::qda()}}.
}
\details{
For \code{\link[MASS:lda]{MASS::lda()}} the result is a matrix whose rows are the quantitative variables in the analysis and whose columns are
the coefficients of linear discriminant functions, \code{LD1}, \code{LD2}, ... This matrix transforms observations to discriminant
functions, normalized so that within groups covariance matrix is spherical.

Quadratic discriminant analysis (from \code{\link[MASS:qda]{MASS::qda()}}) does not produce a simple set of "coefficients" like a linear model.
Instead, the classification is based on parameters that define a quadratic score function for each class, derived from the
estimated class-specific means and covariance matrices. The \code{scaling} components here are an array, where for each group \code{i},
\code{scaling[,,i]} is a matrix which transforms observations so that within-groups covariance matrix is spherical.
}
\examples{
library(MASS)
iris.lda <- lda(Species ~ ., iris)
coef(iris.lda)

iris.qda <- qda(Species ~ ., iris)
coef(iris.qda)

}
