% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\alias{confusion.lda}
\alias{confusion.qda}
\alias{print.confusion}
\title{Confusion matrix for LDA / QDA}
\usage{
confusion(object, ...)

\method{confusion}{lda}(object, ...)

\method{confusion}{qda}(object, ...)

\method{print}{confusion}(x, digits = 2, ...)
}
\arguments{
\item{object}{An object of class \code{lda} or \code{qda}}

\item{...}{Arguments passed to methods}

\item{x}{A \code{confusion} object}

\item{digits}{Number of decimal places for accuracy and error rates}
}
\value{
An object of class \code{c("confusion", "table")}, a table of
frequencies of actual vs. predicted class with attributes \code{accuracy}
and \code{error} giving the overall rates of correct and incorrect prediction.
}
\description{
Calculates a table of the actual vs. predicted class of all observations
in a dataset for objects resulting from \code{\link[MASS:lda]{MASS::lda()}} or \code{\link[MASS:qda]{MASS::qda()}}
}
\examples{

library(MASS)
iris.lda <- lda(Species ~ ., iris)
confusion(iris.lda)

iris.qda <- qda(Species ~ ., iris)
confusion(iris.qda)

}
\seealso{
\code{\link[MASS:lda]{MASS::lda()}}, \code{\link[MASS:qda]{MASS::qda()}}
}
