% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candisc.R
\name{predictor.names}
\alias{predictor.names}
\alias{predictor.names.default}
\title{Get predictor names from a \code{lm}-like model}
\usage{
predictor.names(model, ...)

\method{predictor.names}{default}(model, ...)
}
\arguments{
\item{model}{Model object}

\item{...}{other arguments (ignored)}
}
\value{
A character vector of variable names
}
\description{
Get predictor names from a \code{lm}-like model
}
\section{Methods (by class)}{
\itemize{
\item \code{predictor.names(default)}: \code{"default"} method.

}}
\examples{
#none
}
