\name{candisc-package}
\alias{candisc-package}

\docType{package}
\title{
Generalized Canonical Discriminant Analysis
}
\description{
This package includes functions for computing and visualizing 
generalized canonical
discriminant analyses for a multivariate linear model.
They are designed to provide low-rank visualizations of
terms in a \code{mlm} via the \code{plot} method and the \pkg{heplots} package.
}
\details{
\tabular{ll}{
Package: \tab candisc\cr
Type: \tab Package\cr
Version: \tab 0.5-13\cr
Date: \tab 2008-12-12\cr
License: \tab GPL (>= 2)\cr
}
The organization of functions in this package and the \pkg{heplots} package
may change in a later version.
}

\author{
Michael Friendly and John Fox

Maintainer: Michael Friendly <friendly@yorku.ca>
}
\references{
Gittins, R. (1985). Canonical Analysis: A Review with Applications in Ecology,
Berlin: Springer.


}

\keyword{ package }
\keyword{multivariate}
\seealso{ \code{\link[heplots]{heplot}}

}
%\examples{
%
%}
