% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discrim.R
\name{plot_discrim}
\alias{plot_discrim}
\title{Create a Discriminant Analysis Decision Plot using ggplot.}
\usage{
plot_discrim(
  model,
  vars,
  data = insight::get_data(model),
  resolution = 100,
  point.size = 3,
  showgrid = c("tile", "point", "none"),
  contour = TRUE,
  contour.color = "black",
  tile.alpha = 0.2,
  ellipse = FALSE,
  ...,
  modes.means
)
}
\arguments{
\item{model}{a discriminant analysis model object from \code{MASS::lda()} or \code{MASS::qda()}}

\item{vars}{either a character vector of length 2 of the names of the \code{x} and \code{y} variables, or a formula of form \code{y ~ x}
specifying the axes in the plot.}

\item{data}{data to use for visualization. Should contain all the data needed to use the \code{model} for prediction. The default is to use
the data used to fit the \code{model}.}

\item{resolution}{number of points in x, y variables to use for visualizing the predicted class boundaries and regions.}

\item{point.size}{size of the plot symbols use to show the data observations}

\item{showgrid}{a character string; how to display predicted class regions: \code{"tile"} for \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}, \code{"point"}
for \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}, or \code{"none"} for no grid display.}

\item{contour}{logical (default: \code{TRUE}); should the plot display the boundaries of the classes by contours?}

\item{contour.color}{color of the lines for the contour boundaries (default: \code{"black"})}

\item{tile.alpha}{transparency value for the background tiles of predicted class.}

\item{ellipse}{logical; if \code{TRUE}, 68 percent data ellipses for the groups are added to the plot.}

\item{...}{further parameters passed to \code{predict()}}

\item{modes.means}{levels to use for evaluating predictions using the variables \strong{not} specified in \code{vars}. If not specified,
the function uses the means for quantitative variables, ...}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Discriminant analysis can be more easily understood from plots of the data variables showing how observations are classified.
\code{plot_discrim()} uses the ideas behind \strong{effect plots} (Fox, 1987): Visualize predicted classes of the observations for two focal variables over a
grid of their values, with other variables in a model held fixed. This differs from the usual effect plots in that the predicted
values to be visualized are discrete categories rather than quantitative.

In the case of discriminant analysis, the predicted values are class membership,
so this can be visualized by mapping the categorical predicted class to discrete colors used as the background for the plot, or
plotting the \strong{contours} of predicted class membership as lines (for \verb{[MASS::lda()]}) or auadratic curves (for \verb{[MASS::qda()]}) in the plot.
The predicted class of any observation in the space of the variables displayed can also be rendered as colored \strong{tiles} or \strong{points}
in the background of the plot.
}
\details{
This is an initial version of \code{plot_discrim()}. It is marked "experimental" because the arguments and defaults are still subject to change.

\strong{Customizing colors and shapes}

Since \code{plot_discrim()} returns a \code{"ggplot"} object, you can easily customize colors and shapes by adding scale layers after
the function call. You can also add other graphic layers, such as annotations or labels for the groups.
\itemize{
\item Use \code{scale_color_manual()} \strong{and} \code{scale_fill_manual()} to control the colors used when using \code{showgrid = "tile"}, because that maps
both \strong{both} \code{color} and \code{fill} to the group variable.
\item Use \code{scale_shape_manual()} to control the symbols used for \code{geom_points()}
}
}
\examples{
library(MASS)
library(ggplot2)
library(dplyr)

iris.lda <- lda(Species ~ ., iris)
# formula call: y ~ x
plot_discrim(iris.lda, Petal.Length ~ Petal.Width)

# add data ellipses
plot_discrim(iris.lda, Petal.Length ~ Petal.Width, 
             ellipse = TRUE) 

# without contours
# data ellipses
plot_discrim(iris.lda, Petal.Length ~ Petal.Width, 
             contour = FALSE) 

# specifying `vars` as character names for x, y
plot_discrim(iris.lda, c("Petal.Width", "Petal.Length"))

# Define custom colors and shapes, modify theme() and legend.position
iris.colors <- c("red", "darkgreen", "blue")
iris.pch <- 15:17
plot_discrim(iris.lda, Petal.Length ~ Petal.Width) +
  scale_color_manual(values = iris.colors) +
  scale_fill_manual(values = iris.colors) +
  scale_shape_manual(values = iris.pch) +
  theme_bw(base_size = 14) +
  theme(legend.position = "inside",
        legend.position.inside = c(.8, .25))

# Quadratic discriminant analysis gives quite a different result
iris.qda <- qda(Species ~ ., iris)
plot_discrim(iris.qda, Petal.Length ~ Petal.Width)

}
\references{
Fox, J. (1987). Effect Displays for Generalized Linear Models. In C. C. Clogg (Ed.), \emph{Sociological Methodology}, 1987 (pp. 347–361). Jossey-Bass
}
\seealso{
\code{\link[klaR:partimat]{klaR::partimat()}} for pairwise discriminant plots, but with little control of plot details
}
\author{
Original code by Oliver on SO \url{https://stackoverflow.com/questions/63782598/quadratic-discriminant-analysis-qda-plot-in-r}.

Generalized by Michael Friendly
}
