% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect.R
\name{reflect}
\alias{reflect}
\title{Reflect Columns in an Object, reversing the sign of all elements}
\usage{
reflect(object, columns = 1:2, ...)
}
\arguments{
\item{object}{An object whose columns are to be reflected}

\item{columns}{a vector of indices of the columns to reflect}

\item{...}{Unused}
}
\value{
The object with specified columns of the variable coefficients and observation scores multiplied by -1.
}
\description{
\code{candisc} objects have coefficients for the X and Y and weighted scores for these
whose signs are arbitrary, in the sense that a given column can be reflected (multiplied by -1) without changing the fit.
This function allows one to reflect any columns of the variable coefficients (and corresponding observation scores).
This is often useful for interpreting a biplot, for example when a component (often the first) has all negative signs.
}
\details{
\code{reflect} methods are available for:
\itemize{
\item \code{data.frame}s
\item \code{"cancor"} objects
\item \code{"candisc"} objects
}
}
\author{
Michael Friendly
}
